reformatted code to google conventions

+++ src/View/SimSpecificUI.java
@@ -2,7 +2,6 @@ package View;
 
 import java.util.HashMap;
 import java.util.Map;
-
 import javafx.scene.Group;
 import javafx.scene.control.Label;
 import javafx.scene.control.Spinner;
@@ -11,16 +10,17 @@ import javafx.scene.layout.VBox;
 
 
 public class SimSpecificUI {
+
   private HashMap<String, Spinner> spinners;
 
-  public SimSpecificUI(Group group, Map<String, Object> varNames){
+  public SimSpecificUI(Group group, Map<String, Object> varNames) {
     spinners = new HashMap<>();
     makeSpinners(group, varNames);
   }
 
-  private void makeSpinners(Group group,  Map<String, Object> varNames) {
+  private void makeSpinners(Group group, Map<String, Object> varNames) {
     VBox v = new VBox();
-    for(String k : varNames.keySet()){
+    for (String k : varNames.keySet()) {
       v.getChildren().add(makeInput(k, varNames.get(k)));
       v.setSpacing(20);
     }
@@ -35,11 +35,10 @@ public class SimSpecificUI {
     HBox h = new HBox();
     Label b = new Label(k);
     Spinner s;
-    if(o instanceof Double){
-      s = new Spinner(0,1, (double) o, .05);
-    }
-    else{
-      s = new Spinner(0,100, (int) 0, 1);
+    if (o instanceof Double) {
+      s = new Spinner(0, 1, (double) o, .05);
+    } else {
+      s = new Spinner(0, 100, 0, 1);
     }
     s.setEditable(true);
     spinners.put(k, s);
@@ -48,9 +47,9 @@ public class SimSpecificUI {
     return h;
   }
 
-  public Map<String, Object> getValues(){
-    HashMap <String, Object> ret = new HashMap<>();
-    for (String k : spinners.keySet()){
+  public Map<String, Object> getValues() {
+    HashMap<String, Object> ret = new HashMap<>();
+    for (String k : spinners.keySet()) {
       ret.put(k, spinners.get(k).getValue());
     }
     return ret;
